debian/postrm: include removal of mountpoints in purge
According to the Debian Policy document (chapter on maintainer scripts), the
postrm script is the one that can be called with the purge option passed
(the others don't).

While we are at it, be more verbose in the comments.

Also change the way that tests in the shell scrips are made, to make them
use the internal test command of the shell (thus, removing some potential
forking of other processes) and being easier to read.

The change above was checked with the manpage of dash and also checked with
`checkbashisms`.