debian/pre{inst,rm}: use internal test command of the shell.
Change the way that tests in the shell scrips are made, to make them use the
internal test command of the shell (thus, removing some potential forking of
other processes) and being easier to read.

The change above was checked with the manpage of dash and also checked with
`checkbashisms`.